--- library/GeometriaLab/Acl/Exception/AccessDenied.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace GeometriaLab\Acl\Exception;
-
-class AccessDenied extends \Exception
-{
-
-}
